Rename validate.py
