+++ tests/v110/test_mate.py
@@ -1,6 +1,6 @@
 """Testing breakends (aka BND, mates). Check that they work and that we get the expected output."""
 from utils.beacon_query import call_beacon
-from utils.validate import assert_partly_in, run_test
+from utils.compare import assert_partly_in, run_test
 
 
 @run_test()
